Pipeline周りの整理・リファクタリング
